Implement migrad as an optimization backend
