+++ tensorprob/optimizers/__init__.py
@@ -1,2 +1,3 @@
 from .base import BaseOptimizer
 from .scipy_lbfgsb import ScipyLBFGSBOptimizer
+from .migrad import MigradOptimizer
